refactor: addresses code review feedback on test stability
- Use expect.poll() instead of waitForTimeout(500) for page title verification
- Makes TC-3370 tests more stable and deterministic
- Faster on quick runs, more reliable on slower runs
